testing

+++ skeema/schemas/AdventureWorks/TestTable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `TestTable` (
-  `id` bigint NOT NULL,
+  `id` int NOT NULL,
   `FirstName` varchar(50) NOT NULL,
   `LastName` varchar(50) NOT NULL,
   PRIMARY KEY (`id`)
